requirements.txt
# Core dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0

# Data handling
roboflow>=1.1.0
pyyaml>=6.0
tqdm>=4.65.0

# Metrics and evaluation
scikit-learn>=1.3.0
pandas>=2.0.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utils
python-dotenv>=1.0.0
